Finalized getFValue optimization fix
creates a table structure which contains a 2D array containing the hop
values. This is created when the template is made. The simbox now has a
table array pointer which holds a table for each type of molecule in the
simulation. This is then accessed when calling the getFValue function.