brute-force the recently changes IOUtilities to compile once again
Changes to various structures necessitated changes to IOUtilities, so
as to no longer rely on metroUtil.h and metroUtil.cpp from the previous
file structure. Some of the configuration file parsing is currently
disabled, and some changes were made to StructLibrary.h. Other changes
can be found throughout to shoehorn all this in.