Attached Simulation execution with command parsing
The application can parse the command line arguments, including the
config path, and then create a new Simulation. The simulation creates a
new SerialBox and initializes it with the config scanner class. The
Simulation will fill the Environment settings into the SerialBox and
print it out to the terminal then exit. This is a proof of concept that
the program structure can work.