Including a note on some of the changes made.
As the file says, I made quite a few changes in that last push that
will likely break everything everywhere. But a note to anyone who
happens to need it to work: I had to include .ccp files instead of .h
files, and redefine/alter the definitions of many of the functions we
had running, in Tavis & Nathan’s demo. If there is any question, let me
know. -Albert