Added in default behavior for device detection
I forgot that the program will default to the CPU if no options are
specified (will change in the future). The device querying logic was
terminating the program even when the user specified to run on the CPU
because it was always querying for available devices.

Also I fixed a typo in the help documentation. I was forgetting to print
the application name in the usage string.