Added the Device listing via Command Line Arguments
Added the --list_devices (-Q) command line argument to the program. This
will list all devices that the program can find on the current system
and print out some info about each one. You do not need to specify a
config file when using this command. It will be overriden by the --help
(-h) flag if it is also specified.