fixed one error in IOUtilities caused by failed Git commit, added minimal underpinning for future functionality
Unintentional "attempt" at overloading readInConfig() [as defined in
the IOUtilities header file] rectified; correct return value at this
point should be "void". Also added a few boolean variables to help with
state safety, but they are, as of now, unused; functionality to be
added and used later in the week.