Updated (without testing) IOUtilities with functionality and bug fixes
IOUtilities has more of its methods in place for proper compilation
support, but it is untested and does not conform to expectations
(namely, the Zmatrix filepath must be present and parseable for
anything else to happen; however, if the Zmatrix path is missing, and
if the state file filepath is missing, the program will exit even
earlier)(also, I’m pretty sure things are not properly wired up).
Still, this was an attempt at making fixes for issues Nathan
encountered.