Removed Debug Output and Using Status Interval
I removed all of the debug output that was being generated by the
FileUtilities stuff when parsing the hop tables and atom positions. So
now the simulation should be pretty clean as far as output is concerned.
I also hooked up the Status Interval command line argument with Seth's
code to print out the system energy. So now we can specify how often the
simulation should print out the status updates. If you don't specify
anything, then the default status interval of 100 is used.