Added Results File
At the end of the Simulation we are now saving a results file
(*.results) that has the printed out information about the run. Right
now this is meant to be parsed by scripts that will gather metrics about
lots of runs, but in the future the tool should be able to output a lot
of information as Dr. Acevado was mentioning last meeting.